requirements.txt
slackclient
aiohttp